# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /consultations/request:
    post:
      operationId: createConsultationRequest
      tags: [consultations]
      summary: Patient requests consultation with a doctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationRequestCreate"
      responses:
        "201":
          description: Consultation request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRequest"
  /consultations/{id}/respond:
    put:
      operationId: respondToConsultation
      tags: [consultations]
      summary: Doctor accepts/rejects consultation request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultationResponseRequest"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationRequest"
  /consultations/{id}/payment:
    post:
      operationId: createPayment
      tags: [consultations]
      summary: Create payment for accepted consultation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /consultations/{id}/payment/confirm:
    post:
      operationId: confirmPayment
      tags: [consultations]
      summary: Confirm sandbox payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get consultation details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsultationDetail"
  /consultations/{id}/complete:
    put:
      operationId: completeConsultation
      tags: [consultations]
      summary: Doctor marks consultation complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /consultations/{id}/prescription:
    post:
      operationId: sendPrescription
      tags: [consultations]
      summary: Doctor sends prescription/document to patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionCreate"
      responses:
        "201":
          description: Prescription sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /consultations/{id}/medication:
    post:
      operationId: sendMedication
      tags: [consultations]
      summary: Doctor sends medication order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationCreate"
      responses:
        "201":
          description: Medication order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationOrder"
  /consultations/{id}/medication/{medId}/status:
    put:
      operationId: updateMedicationStatus
      tags: [consultations]
      summary: Update medication delivery status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: medId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationStatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationOrder"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    DoctorProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        registrationNumber:
          type: string
        stateCouncil:
          type: string
        specialty:
          type: string
        subSpecialty:
          type: string
        languages:
          type: array
          items:
            type: string
        regions:
          type: array
          items:
            type: string
        experience:
          type: integer
        education:
          type: string
        bio:
          type: string
        consultationFee:
          type: number
        isOnline:
          type: boolean
        verificationStatus:
          type: string
        symptoms:
          type: array
          items:
            type: string
        rating:
          type: number
        totalConsultations:
          type: integer
      required:
        - id
        - userId
        - name
...
    PatientProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
        language:
          type: string
        region:
          type: string
        bloodGroup:
          type: string
        allergies:
          type: string
        medicalHistory:
          type: string
      required:
        - id
        - userId
        - name
        - email
...
    ConsultationRequestCreate:
      type: object
      properties:
        doctorId:
          type: integer
        symptoms:
          type: string
        notes:
          type: string
      required:
        - doctorId
        - symptoms
    ConsultationResponseRequest:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, rejected]
        reason:
          type: string
      required:
        - status
    ConsultationRequest:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        symptoms:
          type: string
        notes:
          type: string
        status:
          type: string
        reason:
          type: string
        createdAt:
          type: string
        patientName:
          type: string
        patientPhone:
          type: string
        patientLanguage:
          type: string
        patientRegion:
          type: string
        patientMedicalHistory:
          type: string
        patientAllergies:
          type: string
        previousConsultations:
          type: integer
      required:
        - id
        - patientId
        - doctorId
        - symptoms
        - status
...
    ConsultationDetail:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        symptoms:
          type: string
        notes:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        roomId:
          type: string
        createdAt:
          type: string
        completedAt:
          type: string
        doctor:
          $ref: "#/components/schemas/DoctorProfile"
        patient:
          $ref: "#/components/schemas/PatientProfile"
        prescriptions:
          type: array
          items:
            $ref: "#/components/schemas/Prescription"
        medications:
          type: array
          items:
            $ref: "#/components/schemas/MedicationOrder"
      required:
        - id
        - status
    PaymentResponse:
      type: object
      properties:
        consultationId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        sessionId:
          type: string
        message:
          type: string
      required:
        - consultationId
        - amount
        - currency
    PrescriptionCreate:
      type: object
      properties:
        content:
          type: string
        documentUrl:
          type: string
        notes:
          type: string
      required:
        - content
    Prescription:
      type: object
      properties:
        id:
          type: integer
        consultationId:
          type: integer
        content:
          type: string
        documentUrl:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - consultationId
        - content
    MedicationCreate:
      type: object
      properties:
        medications:
          type: array
          items:
            type: string
        instructions:
          type: string
        pharmacyName:
          type: string
      required:
        - medications
    MedicationOrder:
      type: object
      properties:
        id:
          type: integer
        consultationId:
          type: integer
        medications:
          type: array
          items:
            type: string
        instructions:
          type: string
        pharmacyName:
          type: string
        deliveryStatus:
          type: string
          enum: [pending, processing, shipped, delivered]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - consultationId
        - medications
        - deliveryStatus
    MedicationStatusUpdate:
      type: object
      properties:
        deliveryStatus:
          type: string
          enum: [pending, processing, shipped, delivered]
      required:
        - deliveryStatus
...